+++ CHANGELOG.md
@@ -51,6 +51,11 @@ method was called at least once through tests execution
 - Added movements history modules unit tests
 - Allow requesting limits change for all of the account types
 - Update chart tickers
+- `isAccountRoleReseted` state field to the KYC vuex module
+- Checking for previous account role & resetting reason when loading 
+KYC request
+- Checking for reset account role on verification pages
+- `resetReason` field to the `ChangeRoleRequestRecord`
 
 ### Changed
 - Now using new account endpoint for loading the fees